// src/app/shared/models/descuento.model.ts
// Generated by https://quicktype.io

export interface Descuento {
    correlativo?:      string;
    id:                 number;
    idEmpresa:         number;
    idSucursal:         number;
    nombre:      string;
    activo:      boolean;
    idTipoDescuento:        number;
    tipoDescuento?:          string;
    descuentoEstablecido:              number;
    itemsEliminados?:    number[]|null;
    detalle?:            DescuentoDetalle[];
    numeroSucursal?:         number;
}

export interface DescuentoDetalle {
    id?:              number;
    idDescuento?:         number;
    idProducto?:      number;
    codigoProducto?:      string;
    producto?:     string;
    activo?:        boolean;
    descuentoEstablecido?:       number;
    precio?:          number;
    descuento?:        number;
    precioFinal?:           number;

    idTipoDescuento?: number;
    TipoDescuento?:  string;

    idCategoria?:  number;
    categoria?:     string;
}

export interface CambioDescuento {
    codigoProducto:      string;
    descuentoEstablecido:       number;
    descuento:        number;
    idTipoDescuento: number;
}